Added script for image compression
